Separate csv and feather versions for the import scripts
