+++ .gitignore
@@ -1,0 +1,2 @@
+build_neuprint_example/import.report
+build_neuprint_example/import/
