+++ build_neuprint_example/index-mb6.cypher
