+++ build_neuprint_example/neo4j_3_5_import.sh
@@ -2,8 +2,8 @@
 
 UUID=$1
 
-NEO4J_DIR=/neo4j
-IMPORT_BASE_DIR=/neo4j/import
+#NEO4J_DIR=/neo4j
+#IMPORT_BASE_DIR=/neo4j/import
 
 NEO4J_DB=${UUID}_neuprint.db
 
@@ -22,11 +22,12 @@ ROI_NODE_CMD="--nodes=import/Neuprint_ROIs_${UUID}.csv"
 ROI_REL_CMD="--relationships:FoundIn=import/Neuprint_Neuron_ROI_${UUID}.csv --relationships:SynFoundIn=import/Neuprint_SynapseSet_ROI_${UUID}.csv"
 
 date
-cd ${NEO4J_DIR}
+#cd ${NEO4J_DIR}
 echo "................"
 echo "import data"
 echo "................"
 date
 pwd
 # neo4j 3.5 build
-/var/lib/neo4j/bin/neo4j-admin import --database=${NEO4J_DB} ${META_NODE_CMD} ${NEURON_NODE_CMD} ${NEURON_REL_CMD} ${SYNSET_NODE_CMD} ${SYNSET_REL_CMD} ${SYNAPSE_NODE_CMD} ${SYNAPSE_REL_CMD}
+#/usr/bin/neo4j-admin import --database=${NEO4J_DB} ${META_NODE_CMD} ${NEURON_NODE_CMD} ${NEURON_REL_CMD} ${SYNSET_NODE_CMD} ${SYNSET_REL_CMD} ${SYNAPSE_NODE_CMD} ${SYNAPSE_REL_CMD}
+/usr/bin/neo4j-admin import ${META_NODE_CMD} ${NEURON_NODE_CMD} ${NEURON_REL_CMD} ${SYNSET_NODE_CMD} ${SYNSET_REL_CMD} ${SYNAPSE_NODE_CMD} ${SYNAPSE_REL_CMD}
